Reorganized API View Tests
The tests have been split by resource (e.g. basket, order, refund). This should help improve the readability and maintainability of the tests.

XCOM-505
