+++ ecommerce/extensions/api/v2/tests/views/test_payments.py
@@ -1,0 +1,42 @@
+import json
+
+from django.core.cache import cache
+from django.core.urlresolvers import reverse
+from django.test import TestCase, override_settings
+
+from ecommerce.extensions.payment.tests.processors import DummyProcessor, AnotherDummyProcessor
+from ecommerce.tests.mixins import UserMixin
+
+
+class PaymentProcessorListViewTests(TestCase, UserMixin):
+    """ Ensures correct behavior of the payment processors list view."""
+
+    def setUp(self):
+        self.token = self.generate_jwt_token_header(self.create_user())
+
+        # Clear the view cache
+        cache.clear()
+
+    def assert_processor_list_matches(self, expected):
+        """ DRY helper. """
+        response = self.client.get(reverse('api:v2:payment:list_processors'), HTTP_AUTHORIZATION=self.token)
+        self.assertEqual(response.status_code, 200)
+        self.assertSetEqual(set(json.loads(response.content)), set(expected))
+
+    def test_permission(self):
+        """Ensure authentication is required to access the view. """
+        response = self.client.get(reverse('api:v2:payment:list_processors'))
+        self.assertEqual(response.status_code, 401)
+
+    @override_settings(PAYMENT_PROCESSORS=['ecommerce.extensions.payment.tests.processors.DummyProcessor'])
+    def test_get_one(self):
+        """Ensure a single payment processor in settings is handled correctly."""
+        self.assert_processor_list_matches([DummyProcessor.NAME])
+
+    @override_settings(PAYMENT_PROCESSORS=[
+        'ecommerce.extensions.payment.tests.processors.DummyProcessor',
+        'ecommerce.extensions.payment.tests.processors.AnotherDummyProcessor',
+    ])
+    def test_get_many(self):
+        """Ensure multiple processors in settings are handled correctly."""
+        self.assert_processor_list_matches([DummyProcessor.NAME, AnotherDummyProcessor.NAME])
